+++ jax/_src/pallas/mosaic_gpu/lowering.py
@@ -90,6 +90,7 @@ CollectiveAxesType = Sequence[Hashable]
 
 @dataclasses.dataclass(frozen=True, kw_only=True)
 class ResourceEstimatorContext:
+  reduction_scratch_bytes: int
   axis_names: _AxisNames
   lowering_semantics: mgpu.LoweringSemantics
 
@@ -361,7 +362,6 @@ def _run_scoped_resource_estimator(
           f"Unsupported memory space: {aval.memory_space}")
   return rs + _estimate_resources(ctx, jaxpr)
 
-REDUCE_SCRATCH_ELEMS = 128 * 4  # vector of 4 elements per lane in each WG
 
 @_register_resource_estimator(lax.reduce_sum_p)
 @_register_resource_estimator(lax.reduce_max_p)
@@ -370,10 +370,10 @@ def _reduce_resource_estimator(
     ctx: ResourceEstimatorContext, x_aval: jax_core.ShapedArray, *, axes,
     **kwargs
 ) -> Resources:
-  del ctx, axes  # Unused.
+  del x_aval, axes, kwargs  # Unused.
   # We don't need SMEM for some reductions, but it depends on the layout, so we
   # conservatively request the maximum scratch space we might need.
-  return Resources(smem_scratch_bytes=REDUCE_SCRATCH_ELEMS * x_aval.dtype.itemsize)
+  return Resources(smem_scratch_bytes=ctx.reduction_scratch_bytes)
 
 
 @dataclasses.dataclass(frozen=True)
@@ -420,6 +420,8 @@ class ModuleContext:
   mesh_info: pallas_utils.MeshInfo | None
   # See the documentation of unsafe_no_auto_barriers in CompilerParams.
   auto_barriers: bool
+  # See the documentation of reduction_scratch_bytes in CompilerParams.
+  reduction_scratch_bytes: int
   warp_axis_name: str | None = None
 
   @property
@@ -596,8 +598,9 @@ class LoweringRuleContext:
   @property
   def estimator_ctx(self) -> ResourceEstimatorContext:
     return ResourceEstimatorContext(
+        reduction_scratch_bytes=self.module_ctx.reduction_scratch_bytes,
         axis_names=self.module_ctx.axis_names,
-        lowering_semantics=self.module_ctx.lowering_semantics,
+        lowering_semantics=self.module_ctx.lowering_semantics
     )
 
 
@@ -886,6 +889,7 @@ def lower_jaxpr_to_module(
 
   rs = _estimate_resources(
       ResourceEstimatorContext(
+          reduction_scratch_bytes=params.reduction_scratch_bytes,
           axis_names=axis_names, lowering_semantics=lowering_semantics
       ),
       jaxpr,
@@ -982,6 +986,7 @@ def lower_jaxpr_to_module(
         if jax_mesh is not None
         else None,
         auto_barriers=not params.unsafe_no_auto_barriers,
+        reduction_scratch_bytes=params.reduction_scratch_bytes,
     )
     del runtime_smem, grouped_barriers, runtime_barriers
     _ = lower_jaxpr_to_mosaic_gpu(
@@ -2592,12 +2597,11 @@ def _reduce_lowering_rule(op, ctx: LoweringRuleContext, x, *, axes, **kwargs):
         raise NotImplementedError("Multi-axis reductions not supported")
       reduced_dim = x.layout.tiling.tile_dimension(axes[0])
       if any(reduced_dim[d] for d in x.layout.partitioned_warp_dims):
-        size = x.layout.vector_length * 128  # a vector per lane in each WG.
-        if size > REDUCE_SCRATCH_ELEMS:
-          raise NotImplementedError(
-              f"Reduce scratch {size=} exceeds max={REDUCE_SCRATCH_ELEMS}"
-          )
-        scratch_ty = jax.ShapeDtypeStruct(shape=(size,), dtype=x_aval.dtype)
+        dtype_bitwidth = dtypes.itemsize_bits(x_aval.dtype)
+        if dtype_bitwidth % 8:
+          raise NotImplementedError("Sub-byte dtypes not supported")
+        scratch_elems = ctx.module_ctx.reduction_scratch_bytes * 8 // dtype_bitwidth
+        scratch_ty = jax.ShapeDtypeStruct(shape=(scratch_elems,), dtype=x_aval.dtype)
         ctx = ctx.module_ctx.scratch_view(scratch_ty)
       else:
         ctx = contextlib.nullcontext(None)
@@ -2645,7 +2649,9 @@ def _reduce_lowering_rule_wg(
   def i32_attr(value: int) -> ir.IntegerAttr:
     return ir.IntegerAttr.get(ir.IntegerType.get_signless(32), value)
   reduction.attributes["offset"] = i32_attr(ctx.module_ctx.smem_used_bytes)
-  reduction.attributes["scratch_size"] = i32_attr(REDUCE_SCRATCH_ELEMS)
+  # TODO(bchetioui): here, we could just donate all the remaining free SMEM that
+  # we have at this point in time.
+  reduction.attributes["scratch_size"] = i32_attr(ctx.module_ctx.reduction_scratch_bytes)
   return reduction.result